Massive reorganisation and cleanup of code samples
With a week off I've finally had a chance to do a sweep through my Fury Road
code archives to work out what's worth keeping and what I can post up. We ran a
delibrately lean pipe and most of the code ended up being glue and script work
but as a one-shot, we didn't have scope (and it would've been the wrong brief)
to develop facility-style tools.